all comments are corrected in JavaDOC style
