+++ src/main/java/instances/Plant.java
@@ -4,12 +4,12 @@
  * use it for free
  */
 
-package classes;
+package instances;
 
 import interfaces.FarmAble;
 import interfaces.Grow;
 
-abstract class Plant implements Grow {
+public abstract class Plant implements Grow {
 
     private boolean isPlanted;
     private String name;
