+++ src/main/java/instances/Wheat.java
@@ -4,19 +4,18 @@
  * use it for free
  */
 
-package classes;
+package instances;
 
-import interfaces.Consumption;
-import interfaces.FarmAble;
+import interfaces.*;
 
-public final class Corn extends Plant implements Consumption {
+public final class Wheat extends Plant implements Consumption {
 
-    public Corn(String name, float productivity, int harvestTime) {
+    public Wheat(String name, float productivity, int harvestTime) {
         super();
-        setName("Corn " + name);
+        setName("Wheat " + name);
         this.productivity = productivity;
         this.harvestTime = harvestTime;
-        this.price = 2;
+        this.price = 3;
     }
 
     public int getPrice() {
@@ -25,13 +24,12 @@ public final class Corn extends Plant implements Consumption {
 
     @Override
     public void produceHarvest(int time, FarmAble farm) {
-
         if (isPlanted() && growingTime >= harvestTime) {
             int harvest = (int) (productivity * getPlantedArea());
             setPlanted(false);
             setGrowingTime(0);
             setPlantedArea(0);
-            farm.addCorn(harvest);
+            farm.addWheat(harvest);
         }
     }
 
@@ -44,18 +42,35 @@ public final class Corn extends Plant implements Consumption {
     }
 
     @Override
-    public void toConsumption(int time, FarmAble farm) {
-        int money = (int) (time * price * 0.02);
-        farm.addMoney(-money);
+    public int hashCode() {
+        int result = 0;
+        char[] chars = getName().toCharArray();
+        for (char ch : chars
+        ) {
+            result += ch;
+        }
+        return result;
+    }
+
+    @Override
+    public boolean equals(Object o) {
+        return (o instanceof Wheat) && (getName().equals(((Wheat) o).getName()) && (productivity
+                == ((Wheat) o).productivity));
     }
 
     @Override
     public String toString() {
-        return "Corn {"
+        return "Wheat {"
                 + "'" + getName() + '\''
                 + ", productivity=" + productivity
                 + ", harvest time=" + harvestTime
                 + ", price=" + price
                 + '}';
+    }
+
+    @Override
+    public void toConsumption(int time, FarmAble farm) {
+        int money = (int) (time * price * 0.01);
+        farm.addMoney(-money);
     }
 }
